feat(controller): controller 일부 기능(구매 금액 입력, 로또 발행 수량 출력) 구현
